Details.java
To write a java program to creates a class named Employee having the following members : Name, Age, Phone number, Address, Salary. It also has a method named printSalary() which prints the salary of the Employee. Two classes Officer and Manager inherits the Employee class. The Officer and Manager classes have data members specialization and department respectively. Now, assign name, age,
phone number, address and salary to an officer and a manager by making an object of both of these classes and print the same.
(Exercise to understand Inheritance)